Moved Tools to a new Repo
The following tools (which contained a .json file) have moved to this repo:
1. ModelUpdater
2. AdtModelVisualizer
3. OWL2DTDL

https://github.com/Azure/opendigitaltwins-building-tools

The reason this was done, was so that other Ontology providers (e.g. Willow) can make their own REC-based Ontology public and still have the DTDL Validators run well. This was not possible without this move as the non-Ontology related .json files in this repo would get picked up by DTDL Validator, and hence fail the validation.

It also made logical sense that the Tools that are in this repo belong in their own dedicated repo.